Fix bug in prompt.py
There was an edge case where prompt.py could try to get PWD when PWD is
invalid.

An example of such a case would be being in a directory in one
git branch that does not exist in another. Let's call the directory
demo/ and the branches foo and bar respectively. If one was in demo/ on
branch foo and then checked out bar, The prompt would start raising
FileNotFoundErrors while in the now nonexistant directory.